refactored the pledge context into project context
The application data is now imported and utilized in the Project context useReducer hook, to allow for the updates on individual project fields values. In the case of the design, the remaining amount of said project.

Logic is of course handled by the reducer. Looping through the array of projects, and updating the property necessary, if the ids match, between the projects list and the checked against selectedId, coming from the payload.